fix: align job integer types with Postgres and CI API tests
Use i32 for durable job counters in the store so Postgres INTEGER columns
decode correctly, fix serve path signatures for clippy, and run api-integration
without requiring just on CI runners.

+++ crates/paraclete-service/src/observability.rs
@@ -71,7 +71,7 @@ pub mod audit {
         );
     }
 
-    pub fn job_claimed(job_id: uuid::Uuid, attempt_count: i64, worker_id: &str) {
+    pub fn job_claimed(job_id: uuid::Uuid, attempt_count: i32, worker_id: &str) {
         tracing::info!(
             target: "parqonaut_audit",
             event = "job.claimed",
